Revert "v4.0.0"
This reverts commit 2ed5689a730d429660709b6d97448c713a9b0bb5.

This release was created in error. We will be staying on the 3.x
versions to correspond to OpenMRS 3.x, until OpenMRS 4.x.

We are following a "zero-versioning" scheme (where even breaking
changes only merit point releases) until release, at which point
no more breaking changes will be allowed in 3.x.