gracepotma-carbon-doc-suggestions
Draft content for the carbon section of the Dev Guide. Totally open to changes and I know this doesn't include the more technical dev guidance we still need - just thought it would be good to have some starter content (and this is something I could actually help with a little).