Standardise UI shell right panel button behaviour
Ensures that a the icon button for items in the UI shell right panel header changes to a
dismiss button (using the `Close20` icon) when the panel is expanded. This applies to the
App menu and the User menu. In future, this will also apply to the Notifications menu.
Note that this doesn't cover the implementer tools button.