MF-633: Align navbar menu item button interactions (#168)
This PR aligns the interaction behaviour for menu items in the navbar as suggested in Ciaran's
[QA design review doc](https://www.notion.so/Top-Nav-6d97f575ec0d46e0a1fd199b7df6b079). More
specifically, we want that when menu item buttons are first clicked (to open a menu panel or similar),
the menu item button icon gets replaced by an 'X'. Later, when the menu item button gets clicked
(to close a menu panel or similar), the 'X' gets replaced with the original button icon.