Update side-menu-panel.component.tsx
Presently, we have a single slot which is essentially overridden by a route (e.g. /chart) to allow route specific navigation items to be added to the nav bar. This adds a new slot (above the existing page specific slot) which can be used across mf's to add global links to that will always be present independent of the route. Designs are forthcoming for how this navigation will look like but this sets the foundation to allow this to happen.