(fix) Restore missing side menu on tablet (#375)
* (fix) Restore missing side menu on tablet

This PR fixes an issue where the side menu was not appearing on tablet. It does so by changing the referenced extension slot from `nav-menu-slot` to `patient-chart-dashboard-slot`. This change allows the navbar component to show a hamburger menu in the top left hand corner of the page. Clicking on this hamburger menu then displays the side menu.

* Fix broken test