(chore) this fixes the run:shell command
A few dependencies need to be fixed to ensure repeatable builds from the
CLI tool. We have version scripts that are run during the version bump
process to ensure that these versions always match. However, this command
fails if the already-listed version doesn't exist. To deal with this,
we had (temporarily) pinned these versions to the 3.x release. This works,
but it means that when running locally, we're actually pulling in the
old versions and this means that we were (locally) building all the apps
with the 3.x version of @openmrs/webpack-config, which output the apps
in the old format instead of the new format, and hence run:shell appeared
to not be working because none of the apps could be loaded.

Switching the version listed in package.json to "*" solves both of the
above problems. Note that because of our CI processes, these "*"s will
_always_ be replaced with the correct version number.