(fix) Fix module loading error in the translations systems
This PR adds a check to the code in the translations system that handles loading modules. This check ensures that the loaded module exists (i.e. is not `undefined`), before proceeding to run the rest of the code. This fixes an error that currently appears in the devtools console (see screenshot below).