(fix) Fix import map overrides panel
This PR fixes an issue where the overrides list UI in the `import-map-overrides` panel wouldn't display following the recent 4.0 upgrade. It changes the way we install the `import-map-overrides` package in esm-core by switching from installing it as an npm package to the preferred approach - installing it via a `script` tag in our html file.