(fix) Fix error that appears when changing locale
Presently, when a user changes their locale, a error notification appears on the screen. This error appears to be linked to the `getSessionLocation` function defined in `current-user.ts`. Specifically, it appears that this function tries to unsubscribe from an observable whose value is occasionally `undefined`.

This PR moves the subscription's `unsubscribe` call outside of the observable definition, ensuring it doesn't called on an object that might not exist.