(fix) Refactor login experience (#616)
This PR introduces another batch of refactors to the login experience. Mainly, it removes the ChooseLocation component, which renders a loading spinner while fetching a list of login locations. We shouldn't need to show a loading spinner because the LocationPicker already has a loading state. Following this change, when the user logs in, they get immediately redirected to the LocationPicker, which fetches the login locations and renders them in a list. Additionally, the logic that handles redirects now exists in the LocationPicker, eliminating the need for the ChooseLocation component.

* Centralise submit logic and add a submission state