(fix) Webpack now processes esm-framework sanely
There are a few things here, but the key seems to be that Webpack's var
packaging doesn't embed well in other var packaged components. This
commit basically takes us back to using "system" packaging inside the
framework itself. We should explore using proper ES6 Modules for the
libraries or at least the sub-components.