(chore) Extend useConfig stub to handle options object
This commit extends the useConfig stub to accept an optional `options` object which can specify an `externalModuleName`. If the `externalModuleName` is provided, the stub will return a mock config for that external module. Otherwise, it will return the default mock implementation for the current module.
This allows for more flexibility in testing different configurations and modules. More importantly, it should potentially fix a bug where the useConfig stub fails silently when an `externalModuleName` is provided in the options object. See https://github.com/openmrs/openmrs-esm-patient-management/pull/1275 for more information.