(fix) Tweak usage of config property in PageHeader
Tweaks the usage of the config property in PageHeader to make it more robust. More specifically, this commit removes the destructuring of the config object in the component. Because this is a reusable component, the config schema that will be used is that of the consuming frontend module. Thus, the component should not make any assumptions about the shape of the config object. This commit also adds a check for the config object to make sure that the implementationName paragraph is rendered conditionally, based on whether the config object exists.