(refactor) Load styleguide config asynchronously in PageHeader (#1133)
* (refactor) Load styleguide config asynchronously in PageHeader

This PR refactors the PageHeader component to load the styleguide config
asynchronously via the getConfig function. This fixes an issue where the useConfig hook usage would result in the consuming
component failing to render in the test environment when running tests. I'm not sure why this only happens in the test environment,
but it's likely related to the way the styleguide config is loaded in the test environment. It's also possible that the better approach here would be
to refactor the useConfig stub to make it more flexible, but this PR fixes the issue for now. Functionality should be unchanged.

FWIW, the [other use](https://github.com/openmrs/openmrs-esm-patient-management/blob/main/packages/esm-patient-registration-app/src/patient-registration/form-manager.ts#L139-L140) of the styleguide config in the Patient Management app loads the config asynchronously via the getConfig function,
so this PR is consistent with that approach.

* Review feedback