(feat) Disable automatic revalidations by default
These changes to the global SWR config disable automatic revalidations by default. The idea is to set up a conservative caching strategy that minimizes automatic data revalidations. In practice, it is very unlikely that the data in a chart will change frequently, so it makes sense to limit the number of revalidations to a bare minimum.

Note that the SWRConfig component merges the configuration from the parent context. This means that alternative configurations provided further down the line will override the parent values.