(refactor) Switch visit requests to use promises instead of observables
Ports over requests to the Visit API to use promise-based fetch requests instead of observables. Promises are
generally simpler to work with than observables, and this should make the code easier to understand. This change also makes
the code more consistent with other requests to the REST API.

I've also updated the `UpdateVisitPayload` type to be a partial of `NewVisitPayload`, as not all fields are required when
updating a visit. We should only really need to provide the stop datetime when ending a visit, for example.