(chore) Add FetchError interface for more granular error annotation
This PR adds a `FetchError` interface to the framework that can be used to annotate errors that occur during fetch requests. It also amends the `OpenmrsFetchError` class to implement the `FetchError` interface.

Presently, the openmrsFetch hook throws a custom `OpenmrsFetchError` class that includes the response status code and status text if a request fails. This is not always the most helpful message, and it's not always clear what the underlying error was.

The REST API returns a more useful message in the responseBody, and it would be preferable to show this message instead of the status code and status text. With this change, we can then use the FetchError interface further down the line in SWR hooks to annotate errors.