(fix) styleguide: Prevent validateDOMNesting warnings for icon mocks
Presently, we're wrapping stubs for icons and pictograms in a `<div>` tag. This is fine for most cases, but it can
cause a validateDOMNesting warning from React when the parent element for the icon is not a block element.

This commit wraps the stubs in a `span` instead of a `div`. This should prevent the warnings.