(chore) Cache install dependencies step in e2e workflow (#1235)
Makes it so that the `Install dependencies` step in the `run-e2e-on-other-repos` job is cached. Without caching, this step takes a full 2 minutes to run for Patient Chart. The commit also adds working-directory directives to keep the cache keys specific to the active repo. Additionally, these changes makes artifact names and steps more explicit by adding the repo name to both the artifact name and step name.