(fix) Fix edge case with formatting dates in January
In #1280, we added support for estimated birthdates. Unfortunately, the
code there had some checks using !parsed.month to detect null or undefined
values. In addition to treating null and undefined as falsy, JS treats 0
as truthy and hence !0 as falsy. But, for JS Date objects, a month of 0
is January and so on. This meant that we accidentally treated dates in
January as dates without a month.

This PR fixes that but switching from !parsed.month to isNil(parsed.month)
and adds tests to cover the basic functions of the formatPartialDate()
function.