(fix) Setup redirect for static assets served from wrong location
Basically, on some pages (e.g., the chart) its possible to do a hard
refresh that results in trying to load the Carbon's Plex fonts relative
to the current route rather than the base SPA URL. This adds a rewrite
so that these cases serve the right asset.

I couldn't reproduce the issue either on dev3 or using the `develop`
command, so I haven't made any additional changes.