(refactor) Remove unused deceased prop from CustomOverflowMenu
Removes the deceased prop from CustomOverflowMenu and its associated styles. The prop was originally added to the patient banner component
to support displaying the `deceased` variant of the patient header https://zeroheight.com/23a080e38/p/6663f3-patient-header/t/page-6663f3-57593405-17. It looks like
the prop was mistakenly added to the CustomOverflowMenu component, where it is not used.