(fix) Restore name prop for user menu button (#1388)
This fixes an issue affecting the user onboarding app where the user menu button was not getting triggered correctly.
This issue manifested itself in a failing e2e test after updating the versions of the core tooling and
framework used in the project. More specifically, this fix restores the name prop for the user menu button,
which was removed in a previous commit. It also adds a ts-ignore comment to the user menu button component to
avoid type errors.